Set sane lowpass value and expose EQ settings
The lowpass value should be modified in system.c but the libretro version never gets there for some reason, so we had to go ahead and subtract the value here instead. I think the same thing is happening with the EQ, actually, because the typical default value of 1 doesn't really make sense when 100 seems to be the full volume.